jungle_chat_py/requirements.txt
# Python >= 3.11
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-dotenv>=0.19.0,<1.0.0
pydantic>=2.0.0

# AI和向量数据库相关
google-generativeai>=0.3.0
pinecone-client>=2.2.1
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.2.38
langchain-google-genai>=0.0.3
langchain-huggingface>=0.0.6
sentence-transformers>=2.2.2

# 数据处理
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0 